Fully import ‎Skip and SkipWhile tests
Convert ‎Skip and SkipWhile  tests imported in #2874 into a more directly
XUnit-ian style

Split existing tests between Skip and SkipWhile.

Of Skip tests, the following were omitted as duplicating
an existing test:

Skip1.Test1() → SkipExcessive()
Skip2.Test2() → SkipNone()
Skip3.Test3() → SkipNegative()
Skip5.Test5() → SkipAllExactly()

Of the SkipWhile tests SkipWhile1a.Test1a() and SkipWhile1b.Test1b() were
omitted as duplicating SkipWhileAllTrue()

SkipWhile6.Test6() was crippled in the original, probably because it causes an
OutOfMemoryException. Fixed this, but it was very slow. Improved the
running time, but it's still very slow so marked it as Skipped.
